Fix warning up to Wconversion

+++ evaluation/performance/line_fit.cpp
@@ -18,7 +18,7 @@ constexpr float th_low = 0.004f, th_high = 0.012f;
 
 struct FitPerformaceData : public TaskData {
   FitPerformaceData(const std::string& n, const cv::Mat& s)
-      : TaskData(n, s), nms(th_low, th_high), edge(10, 3, 3, grad.magnitudeThreshold(th_low)) {
+      : TaskData(n, s), nms(th_low, th_high), edge(10, 3, 3, static_cast<float>(grad.magnitudeThreshold(th_low))) {
     if (src.channels() == 3) cv::cvtColor(src, src, cv::COLOR_BGR2GRAY);
     grad.process(src);
     nms.process(grad);
