+++ evaluation/thesis/gradient_orientation.cpp
@@ -339,8 +339,8 @@ int main() {
       new GradientOdd<PCSqf<uchar, double>>({NV("grad_scale", 1), NV("grad_muls", 3), NV("grad_kernel_spacing", 2.5)}),
       "PCF Po"));
 
-  int cols = gradI.size() + gradF.size() + 1;
-  int rows = 7;
+  size_t cols = gradI.size() + gradF.size() + 1;
+  size_t rows = 7;
   std::vector<std::vector<std::string>> table_box, table_disk;
   table_box.resize(cols);
   table_disk.resize(cols);
@@ -365,7 +365,7 @@ int main() {
   cv::Mat disk_mask = sobel.magnitude() > sobel.magnitudeThreshold(magTh);
 
 
-  int row = 1;
+  size_t row = 1;
   for_each(gradI.begin(), gradI.end(), [&](Entry<short>& e) {
     table_box[row][0] = e.name;
     table_disk[row][0] = e.name;
