+++ examples/imgproc/pyramid_test.cpp
@@ -58,13 +58,13 @@ int main(int argc, char** argv) {
   int64 start = cv::getTickCount();
   double t{};
   for (std::size_t i = 0; i != imgP.size(); ++i) {
-    sobel.process(imgP[i]);
-    imgGrad[i] = sobel.magnitude();
+    sobel.process(imgP[static_cast<int>(i)]);
+    imgGrad[static_cast<int>(i)] = sobel.magnitude();
     nms.process(sobel);
-    imgNMS[i] = nms.hysteresis();
-    if (i == 0) t = (cv::getTickCount() - start) * 1000.0 / cv::getTickFrequency();
+    imgNMS[static_cast<int>(i)] = nms.hysteresis();
+    if (i == 0) t = static_cast<double>(cv::getTickCount() - start) * 1000.0 / cv::getTickFrequency();
   }
-  double time = (cv::getTickCount() - start) * 1000.0 / cv::getTickFrequency();
+  double time = static_cast<double>(cv::getTickCount() - start) * 1000.0 / cv::getTickFrequency();
   std::cout << "time image: " << t << "ms, time pyramid: " << time << "ms" << std::endl;
   showPyramid("img", imgP);
   showPyramid("grad", imgGrad);
